Implement announcement banner on homepage
An announcement banner has been added to the homepage with a link to further detail. The banner is designed to be dynamic and responsive, with CSS configurations included for fitting to various screen sizes. A 'hide' button has also been provided to allow users to remove the banner once read.